Split planner.py
